hid_t cast to int and -lhdf5 added in link step
On some systems hid_t is not the right size for %d in printf, so it must
be cast. Also, I don't know why my mkoctfile is content with me leaving
out -lhdf5...

+++ h5file.cpp
@@ -159,7 +159,7 @@ int H5File::select_hyperslab(const Matrix& start, const Matrix& count,
     if (h5_dims[i] < end)
     {
       error("In dimension %d, dataset only has %d elements, but at least %d"
-            " required for requested hyperslab", i+1, h5_dims[i], end);
+            " required for requested hyperslab", i+1, (int)h5_dims[i], end);
       return 1;
     }
   }
